requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
py-clob-client>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
ccxt>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0
schedule>=1.2.0
google-auth>=2.28.0
firebase-admin>=6.4.0
